+++ src/java.desktop/unix/classes/sun/java2d/vulkan/WLVKSurfaceData.java
@@ -60,7 +60,8 @@ public abstract class WLVKSurfaceData extends VKSurfaceData implements WLSurface
         final int backgroundRGB = peer.getBackground() != null
                 ? peer.getBackground().getRGB()
                 : 0;
-        int scale = ((WLGraphicsConfig)peer.getGraphicsConfiguration()).getWlScale();
+        // TODO: support fractional scaling
+        int scale = (int)(((WLGraphicsConfig)peer.getGraphicsConfiguration()).getWlScale());
         initOps(peer.getBufferWidth(), peer.getBufferHeight(), scale, backgroundRGB);
     }
